# backend-1/src/open-api/swagger.yaml
openapi: 3.0.0
info:
  title: IoT API
  version: 1.0.0
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/configuration:
    get:
      summary: Get device configuration
      description: >-
        Returns complete device configuration or filtered by topic/subtopic.
        Data is cached in Redis.
      tags:
        - Configuration
      parameters:
        - in: query
          name: topic
          schema:
            type: string
          description: Configuration topic (e.g., show-all, mqtt)
        - in: query
          name: subtopic
          schema:
            type: string
          description: Configuration subtopic (e.g., show-all)
      responses:
        '200':
          description: Configuration data
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Internal server error
    post:
      summary: Create/Update configuration
      description: Creates or updates device configuration. Requires authentication.
      tags:
        - Configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Configuration updated successfully
        '500':
          description: Failed to update configuration
    patch:
      summary: Partially update configuration
      description: Partially updates device configuration. Requires authentication.
      tags:
        - Configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Configuration updated successfully
        '404':
          description: Configuration not found
        '500':
          description: Failed to update configuration
...
  /api/device/port/{portId}:
    get:
      summary: Get specific port information
      description: >-
        Returns information for a specific port. Data is cached in Redis for
        performance.
      tags:
        - Device Ports
      parameters:
        - in: path
          name: portId
          required: true
          schema:
            type: string
          description: Port identifier (e.g., XD1, XF1, X0, HMI)
      responses:
        '200':
          description: Port information array with fields and data
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - type: array
                      items:
                        type: string
                    - type: object
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  detail:
                    type: string
                  type:
                    type: string
                  status:
                    type: integer
        '404':
          description: Port not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  detail:
                    type: string
                  type:
                    type: string
                  status:
                    type: integer
        '500':
          description: Internal server error
  /api/device/ports:
    get:
      summary: Get all ports information
      description: Returns information for all device ports. Data is cached in Redis.
      tags:
        - Device Ports
      responses:
        '200':
          description: Object containing all ports data
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...